requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
Flask_CKEditor==0.4.6
flask_sqlalchemy==3.0.5
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0